Make Agora compile locally, add steps to README
